Reland "Prevent unsafe narrowing: base/ for Linux part 2 of n"
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758876
Owners-Override: danakj <danakj@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024348}

+++ base/profiler/stack_copier_signal.cc
@@ -37,7 +37,7 @@ class AsyncSafeWaitableEvent {
     // futex() can wake up spuriously if this memory address was previously used
     // for a pthread mutex. So, also check the condition.
     while (true) {
-      int res =
+      long res =
           syscall(SYS_futex, futex_int_ptr(), FUTEX_WAIT | FUTEX_PRIVATE_FLAG,
                   0, nullptr, nullptr, 0);
       if (futex_.load(std::memory_order_acquire) != 0)
